Updated calls to DblToString() to set limitDecimalsForLargeValues to true for intensities

+++ DeconTools.Backend/ProcessingTasks/UIMF_TICExtractor.cs
@@ -16,7 +16,7 @@ namespace DeconTools.Backend.ProcessingTasks
 
             UIMFRun uimfRun = (UIMFRun)resultList.Run;
             float ticValue = uimfRun.GetTIC(uimfRun.CurrentScanSet.PrimaryScanNumber, uimfRun.CurrentIMSScanSet.PrimaryScanNumber);
-            uimfRun.CurrentIMSScanSet.TICValue = (float)ticValue;
+            uimfRun.CurrentIMSScanSet.TICValue = ticValue;
         }
     }
 }
